Move use.hpp out of include/deepstream
