+++ src/client.cpp
@@ -18,7 +18,7 @@
 
 #include <deepstream/client.hpp>
 #include <deepstream/message.hpp>
-#include <deepstream/use.hpp>
+#include "use.hpp"
 
 #include <cassert>
