+++ test/random.cpp
@@ -20,7 +20,7 @@
 #include <deepstream/buffer.hpp>
 #include <deepstream/message_builder.hpp>
 #include <deepstream/random.hpp>
-#include <deepstream/use.hpp>
+#include "../src/use.hpp"
 
 namespace deepstream {
 namespace random {
